requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-magic>=0.4.27
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0

langchain==0.3.9
langgraph==0.2.56
langchain-openai>=0.0.8
langchain-core~=0.3.60

datasets==3.6.0
pyarrow==20.0.0
python-multipart>=0.0.6

kagglehub==0.3.12
pandas~=2.2.3
typing_extensions~=4.13.2
torch>=2.0.0
transformers>=4.0.0
captum>=0.6.0